Remove all refcat runs when regenerating refcats.
The original implementation of generate_refcats_gen3.py was written
when there was no guarantee of the preloaded repository having any
particular state, and was made as conservative as possible to allow
by-hand verification of the new refcats. Now that scripted ap_verify
dataset management is routine, this is unnecessary complexity.